refactor(effect): rename effect-island to effect, share runtime
- Move src/main/effect-island/* to src/main/effect/* (preserves git blame via git mv).

- Refactor bash runtime.ts to delegate singleton storage to the shared _shared/runtime.ts registry while keeping its public exports (getRuntime, setMainWindow, disposeRuntime) unchanged.

- Switch app shutdown from bash-specific disposeRuntime() to disposeAllRuntimes() so future islands tear down without extra wiring.

- Update bash-handlers.ts and main/index.ts import paths.

- Update vitest.workspace.ts main include glob.

bashService facade signatures unchanged; no callers migrated.
