+++ src/main/effect/bash/__tests__/bash-service.effect.test.ts
