+++ src/main/effect/bash/errors.ts
