+++ src/main/effect/bash/facade.ts
