+++ src/main/effect/bash/runtime.ts
@@ -1,0 +1,34 @@
+import type { BrowserWindow } from 'electron'
+import { ManagedRuntime } from 'effect'
+
+import { Bash } from './service'
+import { AppLive } from './layers'
+import {
+  getOrCreateRuntime,
+  disposeRuntime as disposeRuntimeShared
+} from '../_shared/runtime'
+
+const ISLAND_NAME = 'bash'
+
+/**
+ * Window reference is mutable - it's set after the BrowserWindow is created
+ * (see registerBashHandlers in src/main/ipc/bash-handlers.ts) and is read
+ * lazily by EventSinkLive when streaming events back to the renderer. The
+ * ref object itself is captured at runtime construction time and survives
+ * across calls.
+ */
+const windowRef: { current: BrowserWindow | null } = { current: null }
+
+export const getRuntime = (): ManagedRuntime.ManagedRuntime<Bash, never> =>
+  getOrCreateRuntime(ISLAND_NAME, () => ManagedRuntime.make(AppLive(windowRef)))
+
+export const setMainWindow = (win: BrowserWindow): void => {
+  windowRef.current = win
+}
+
+/**
+ * Dispose this island's runtime. Equivalent to (and delegates to)
+ * `disposeRuntime('bash')` from the shared registry. Most callers should
+ * prefer `disposeAllRuntimes()` from `../_shared/runtime`.
+ */
+export const disposeRuntime = (): Promise<void> => disposeRuntimeShared(ISLAND_NAME)
