+++ src/main/effect/bash/service.ts
