+++ src/main/effect/bash/types.ts
