+++ src/main/index.ts
@@ -61,8 +61,8 @@ import { perfDiagnostics } from './services/perf-diagnostics'
 import { configure as configureCodexDebugLogger } from './services/codex-debug-logger'
 import { ptyService } from './services/pty-service'
 import { scriptRunner } from './services/script-runner'
-import { bashService } from './effect-island/bash/facade'
-import { disposeRuntime } from './effect-island/bash/runtime'
+import { bashService } from './effect/bash/facade'
+import { disposeAllRuntimes } from './effect/_shared/runtime'
 import { registerBashHandlers } from './ipc/bash-handlers'
 import { registerTicketImportHandlers } from './ipc/ticket-import-handlers'
 import { initTicketProviderManager, GitHubProvider, JiraProvider } from './services/ticket-providers'
@@ -764,7 +764,7 @@ app.on('will-quit', async () => {
   cleanupScripts()
   // Cleanup running bash runs (best-effort, no await)
   bashService.killAll()
-  await disposeRuntime()
+  await disposeAllRuntimes()
   // Release any held power save blocker so the display can sleep again
   cleanupPowerSaveBlocker()
   // Cleanup file tree watchers
