+++ vitest.workspace.ts
@@ -37,7 +37,7 @@ export default defineWorkspace([
         'test/kanban/session-3/**/*.test.ts',
         'test/codex-migration/**/*.test.ts',
         'test/utils/**/*.test.ts',
-        'src/main/effect-island/**/*.test.ts'
+        'src/main/effect/**/*.test.ts'
       ],
       globals: true
     },
